Make code cleaner
